+++ zippyipscanner/forms/uiaboutdialog.py
@@ -1,0 +1,65 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file 'aboutdialog.ui'
+#
+# Created by: PyQt5 UI code generator 5.10
+#
+# WARNING! All changes made in this file will be lost!
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+class Ui_AboutDialog(object):
+    def setupUi(self, AboutDialog):
+        AboutDialog.setObjectName("AboutDialog")
+        AboutDialog.resize(640, 250)
+        self.gridLayout = QtWidgets.QGridLayout(AboutDialog)
+        self.gridLayout.setObjectName("gridLayout")
+        self.label_5 = QtWidgets.QLabel(AboutDialog)
+        self.label_5.setObjectName("label_5")
+        self.gridLayout.addWidget(self.label_5, 3, 1, 1, 1)
+        spacerItem = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+        self.gridLayout.addItem(spacerItem, 4, 1, 1, 1)
+        self.label = QtWidgets.QLabel(AboutDialog)
+        self.label.setObjectName("label")
+        self.gridLayout.addWidget(self.label, 0, 0, 1, 1)
+        self.label_2 = QtWidgets.QLabel(AboutDialog)
+        self.label_2.setObjectName("label_2")
+        self.gridLayout.addWidget(self.label_2, 2, 0, 1, 1)
+        self.label_4 = QtWidgets.QLabel(AboutDialog)
+        self.label_4.setObjectName("label_4")
+        self.gridLayout.addWidget(self.label_4, 3, 0, 1, 1)
+        self.buttonBox = QtWidgets.QDialogButtonBox(AboutDialog)
+        self.buttonBox.setOrientation(QtCore.Qt.Horizontal)
+        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Close)
+        self.buttonBox.setObjectName("buttonBox")
+        self.gridLayout.addWidget(self.buttonBox, 5, 1, 1, 1)
+        self.label_3 = QtWidgets.QLabel(AboutDialog)
+        self.label_3.setObjectName("label_3")
+        self.gridLayout.addWidget(self.label_3, 2, 1, 1, 1)
+        self.label_6 = QtWidgets.QLabel(AboutDialog)
+        self.label_6.setObjectName("label_6")
+        self.gridLayout.addWidget(self.label_6, 1, 0, 1, 1)
+        self.label_7 = QtWidgets.QLabel(AboutDialog)
+        self.label_7.setObjectName("label_7")
+        self.gridLayout.addWidget(self.label_7, 1, 1, 1, 1)
+        self.labelVersion = QtWidgets.QLabel(AboutDialog)
+        self.labelVersion.setObjectName("labelVersion")
+        self.gridLayout.addWidget(self.labelVersion, 0, 1, 1, 1)
+
+        self.retranslateUi(AboutDialog)
+        self.buttonBox.accepted.connect(AboutDialog.accept)
+        self.buttonBox.rejected.connect(AboutDialog.reject)
+        QtCore.QMetaObject.connectSlotsByName(AboutDialog)
+
+    def retranslateUi(self, AboutDialog):
+        _translate = QtCore.QCoreApplication.translate
+        AboutDialog.setWindowTitle(_translate("AboutDialog", "About Buildertron..."))
+        self.label_5.setText(_translate("AboutDialog", "https://github.com/swprojects/Buildertron"))
+        self.label.setText(_translate("AboutDialog", "Buildertron"))
+        self.label_2.setText(_translate("AboutDialog", "Author"))
+        self.label_4.setText(_translate("AboutDialog", "Homepage"))
+        self.label_3.setText(_translate("AboutDialog", "Simon Wu"))
+        self.label_6.setText(_translate("AboutDialog", "License"))
+        self.label_7.setText(_translate("AboutDialog", "The MIT License - https://opensource.org/licenses/MIT"))
+        self.labelVersion.setText(_translate("AboutDialog", "v0.0"))
+