+++ HISTORY.rst
@@ -3,4 +3,19 @@
 History
 -------
 
-Pre-release
+1.0.2
+    - Implement filtering of alive results
+    - Added settings dialog
+    - Added shortcut keys for scanning and other dialogs
+    - Added clear scan history
+    - Added clear results button
+    - Added disable splashscreen
+    - Use generated UI forms from Qt Creator
+    - Enabled HighDPI scaling
+    - Re-add check for updates
+    - Fix custom scan parsing
+1.0.1
+    - Ported to pyqt
+    - Remove check for updates
+0.1
+    - First release